refactor(mcp): extract deploy_apply tool module
