+++ openspec/changes/refactor-mcp-deploy-apply-tool-module/README.md
@@ -1,0 +1,3 @@
+# refactor-mcp-deploy-apply-tool-module
+
+Refactor: extract MCP deploy_apply tool implementation into a dedicated module
