+++ openspec/changes/refactor-mcp-deploy-apply-tool-module/proposal.md
@@ -1,0 +1,14 @@
+# Change: Extract MCP deploy_apply tool implementation into its own module
+
+## Why
+`src/mcp/tools.rs` contains implementations for many tools and is large. Extracting the `deploy_apply` tool handler into a dedicated module improves readability and makes future changes safer and easier to review.
+
+## What Changes
+- Move the `deploy_apply` tool implementation from `src/mcp/tools.rs` into `src/mcp/tools/deploy_apply.rs`.
+- Keep the public MCP surface and JSON output unchanged.
+
+## Impact
+- Affected specs: `agentpack-mcp` (no behavioral change expected).
+- Affected code:
+  - `src/mcp/tools.rs`
+  - `src/mcp/tools/deploy_apply.rs`
